+++ src/printer.h
@@ -1,0 +1,8 @@
+#ifndef PRINTER_H
+#define PRINTER_H
+
+void printer(const char* msg);
+void printer_closing(const char* msg);
+void printer_header(const char* command);
+
+#endif  // PRINTER_H